Store step/stage result as attribute on corresponding spans
Unlike otel.status_code, which can only be OK or ERROR, the step result can also be UNSTABLE, NOT_BUILT (for skipped steps), ABORTED, etc.

The result of a stage is calculated from the error and warning actions attached to the FlowNodes between the stage's StepStartNode and StepEndNode, using org.jenkinsci.plugins.workflow.pipelinegraphanalysis.StatusAndTiming (which is the implementation that ultimately also provides stage results for BlueOcean).