173 Fix override from setSpanAttribute not working
The problem is that notifyOfNewStep is executed after the spanAttribute
step, thus any attribute set by (set/with)SpanAttribute is reverted by
attributes from the context with the same key. This occurs because
(set/with)SpanAttribute steps do not have their own span, so the context
attributes are applied to some spans several times.