Change `assumeTrue` import (#1030)
I'm working on getting `opentelemetry` into the plugin BOM. The one item that is holding me back is the failing tests when trying to start Elasticsearch. This is because there is no Docker available in the BOM build infra.

I'm running into an issue when using the JUnit5 `Assumptions.assumeTrue` where it throws an exception:

```
[ERROR] Tests run: 1, Failures: 0, Errors: 1, Skipped: 0, Time elapsed: 9.615 s <<< FAILURE! -- in io.jenkins.plugins.opentelemetry.backend.elastic.ElasticsearchBackendITTest
[ERROR] io.jenkins.plugins.opentelemetry.backend.elastic.ElasticsearchBackendITTest -- Time elapsed: 9.615 s <<< ERROR!
org.opentest4j.TestAbortedException: Assumption failed: assumption is not true
        at io.jenkins.plugins.opentelemetry.rules.CheckIsDockerAvailable.before(CheckIsDockerAvailable.java:19)
        at org.jvnet.hudson.test.JenkinsRule$1.evaluate(JenkinsRule.java:659)
        at java.base/java.util.concurrent.FutureTask.run(FutureTask.java:317)
        at java.base/java.lang.Thread.run(Thread.java:1583)
```

instead of just skipping the test.

By changing from JUnit5 `Assumptions` to JUnit4 `Assume`, the tests skip as expected.

If you accept this PR, we'll be able to get `opentelemetry` into plugin BOM.