Use plugin BOM more widely in pom file (#1131)
* Rely on plugin BOM for version of Job DSL plugin

Confirmed with `mvn dependency:tree` that the version provided by the
plugin BOM is the same as the version that was previously declared in
the pom file.

* Rely on plugin BOM for ant version number

Confirmed with mvn dependency:tree that same version is already
provided by plugin BOM

* Rely on plugin BOM for maven plugin version

Confirmed with `mvn dependeency:tree` that it is the same version

* Rely on plugin BOM for Jackson 2 API version

Jackson 2 API plugin 2.19.0-404.x has been suspended from distribution
because it broke the Jenkins kubernetes client plugin in some
important cases.

https://github.com/jenkins-infra/update-center2/pull/869 has more
details.

* Rely on plugin BOM for httpcomponents 5 API version

Confirmed with `mvn dependency:tree` that the version remains the same